Move all logic from general manager to stat tracker, delete general manager
